% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{broken_stick}
\alias{broken_stick}
\title{Broken-stick Distribution for 'p' Pieces.}
\usage{
broken_stick(p)
}
\arguments{
\item{p}{An interger that specify the number of pieces}
}
\value{
broken-stick distribution for `p` pieces.
}
\description{
Compute the expected values of the broken-stick distribution for `p` pieces.
}
\examples{
broken_stick_out_20 = broken_stick(20)
}
