% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_study}
\alias{check_study}
\title{Checking data study properties}
\usage{
check_study(study)
}
\arguments{
\item{study}{The study to be checked.}
}
\description{
This function checks study properties.
}

