% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_survival_table}
\alias{compute_survival_table}
\title{A Function That Computes Survival Table.}
\usage{
compute_survival_table(probe_names, sample_names, exp_grp_key, study,
  suffix = "", USE_CACHE = FALSE, PLOT_SCURVE = FALSE)
}
\arguments{
\item{probe_names}{index of probes to consider}

\item{sample_names}{index of samples to consider}

\item{exp_grp_key}{the column name of the exprimental grouping to considere}

\item{study}{An Reference classes object that contains data and metadata}

\item{suffix}{A character string used to suffix the cache file nam}

\item{USE_CACHE}{A boolean set to TRUE if cache could be used}

\item{PLOT_SCURVE}{A boolean set to TRUE if Survival curve of the corresponding `exp_grp_key` needs to be ploted.}
}
\value{
A dataframe taht contains survival infirmations
}
\description{
This function computes the false discovery rate from a vector of independent p-values.
It extracts the cutoff corresponding to the specified FDR. See Benjamini & Hochberg 1995 paper.
}

