% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{create_study}
\alias{create_study}
\title{Study Factory}
\usage{
create_study(cache_filename)
}
\arguments{
\item{cache_filename}{A character string that describes the study cache file name.}
}
\value{
It returns an object of the Study class.
}
\description{
Intanciates an object of the epimedtools internal class Study.
}
\details{
This function intanciates an object of the epimedtools internal class Study. You can learn more about Study class by typing ?epimedtools::Study
}
\examples{
# create study and load data from GSE57831_series_matrix.txt.gz contained in the package
study = create_study()
study$series_matrix_filename = system.file(
   "extdata/GSE26471", 
   "GSE26471_series_matrix.txt.gz", 
   package = "epimedtools"
 )  
# str(study$get_gset(dest_dir="data"))

# load data from GEO
study = create_study()
study$gse = "GSE26471" # 1 sample, expression
# study$gse = "GSE42707" # 1 sample
# study$gse = "GSE56382" # 2 samples
# study$gse = "GSE49585" # 2 samples
# str(study$get_gset(dest_dir="data"))

# cache study
study$cache_it("/tmp/tmp_cached_study.rds")
# load a cached study
study = create_study("/tmp/tmp_cached_study.rds")
}

