% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discr}
\alias{discr}
\title{A Function Discretize a Vector of Numeric}
\usage{
discr(v, nd = 5, breaks)
}
\arguments{
\item{v}{A vector of numeric.}

\item{nd}{The number of classes.}

\item{breaks}{An increasing vector of numeric.}
}
\value{
A vector of factors.
}
\description{
Takes a vector v of numeric. Replaces the values of v by integers
in (1:nd), according to values of breaks (increasing vector of numeric).
If breaks is missing, nd regular quantiles are taken.
}

