% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_gse_raw_tar}
\alias{download_gse_raw_tar}
\title{Retrieving _RAW.tar archive from GEO.}
\usage{
download_gse_raw_tar(gse, datashare_dir = "~/projects/datashare")
}
\arguments{
\item{gse}{A GSE id}

\item{datashare_dir}{A string describing the targeted directory.}
}
\description{
This function retrieves _RAW.tar archgive from GEO to the file system.
}
