% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_compute_matrix_scale}
\alias{dt_compute_matrix_scale}
\title{Binding of deepTools `computeMatrix` Function.}
\usage{
dt_compute_matrix_scale(regions_filename, score_filename, out_filename,
  bin_size = 50, before_region_start_length = 0,
  after_region_start_length = 0, region_body_length,
  number_of_processors = 12, blacklist_filename = NULL,
  FORCE_EXEC = FALSE)
}
\arguments{
\item{regions_filename}{deepTools `computeMatrix` parameter.}

\item{score_filename}{deepTools `computeMatrix` parameter.}

\item{out_filename}{deepTools `computeMatrix` parameter.}

\item{bin_size}{deepTools `computeMatrix` parameter.}

\item{before_region_start_length}{deepTools `computeMatrix` parameter.}

\item{after_region_start_length}{deepTools `computeMatrix` parameter.}

\item{region_body_length}{deepTools `computeMatrix` parameter.}

\item{number_of_processors}{deepTools `computeMatrix` parameter.}

\item{blacklist_filename}{deepTools `computeMatrix` parameter.}

\item{FORCE_EXEC}{a boolean that force call to deepTools `computeMatrix` function.}
}
\description{
This function binds deepTools `computeMatrix` function.
}

