% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimedtools.R
\docType{package}
\name{epimedtools}
\alias{epimedtools}
\alias{epimedtools-package}
\title{epimedtools: A package for computating the notorious bar statistic.}
\description{
The epimedtools package provides three categories of important functions:
epimedtools, bar and baz.
}
\section{epimedtools functions}{

The epimedtools functions ...
}

