% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{et_pairs}
\alias{et_pairs}
\title{A Function that Plots Pairs Correpaltion}
\usage{
et_pairs(...)
}
\arguments{
\item{...}{Parameters passed to pairs function.}
}
\description{
This function enhances pairs fucntion by adding decoration in classic pairs graphic.
}
