% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FDR}
\alias{FDR}
\title{A Function That Computes de False Discovery Rate.}
\usage{
FDR(x, FDR)
}
\arguments{
\item{x}{A vector of independent p-values.}

\item{FDR}{The false discovery rate.}
}
\value{
The corresponding cutoff.
}
\description{
This function computes the false discovery rate from a vector of independent p-values. 
It extracts the cutoff corresponding to the specified FDR. See Benjamini & Hochberg 1995 paper.
}

