% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fuse_exp_grp}
\alias{fuse_exp_grp}
\title{Fuse two experimental grouping}
\usage{
fuse_exp_grp(exp_grp1, exp_grp2, by = "row.names")
}
\arguments{
\item{exp_grp1}{The first experimental grouping.}

\item{exp_grp2}{The second experimental grouping.}

\item{by}{Specifications of the columns used for merging. See 'Details' in ?merge.}
}
\value{
The fused experimental grouping
}
\description{
This function tales two experimental grouping as parameter and merge them according to their row names.
}
