% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gdc_client_wrapper}
\alias{gdc_client_wrapper}
\title{gdc-client wrapperChecking data study properties}
\usage{
gdc_client_wrapper(manifest_filename, dest_dir, files)
}
\arguments{
\item{manifest_filename}{the path to the manifest file}

\item{dest_dir}{the directory to put data}
}
\description{
This function wraps gdc-client.
}

