% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprin.R
\name{get_exprin_app}
\alias{get_exprin_app}
\title{A Function That runs Heatmap App}
\usage{
get_exprin_app(studies, gene_symbols, height = "5000", var = c("samples",
  "raw", "genome"), USE_CLUST = c(TRUE, FALSE), kern_dim = 1,
  ALL_CURVE = FALSE, nb_clust = 3, pf_col_label = "exo",
  exp_grp_col = "tissue_status", exp_grp_col_label, second,
  BREAK_TOO_EXPENSIVE = BREAK_TOO_EXPENSIVE, ...)
}
\arguments{
\item{height}{string specifying the height oh the app}

\item{var}{A vector defining the distances used to cluster data}

\item{USE_CLUST}{A boolean specifying if dendogram needs to be plotted}

\item{kern_dim}{the size of the kernel used to smooth RNAseq signal}

\item{ALL_CURVE}{A boolean specifying if individual RNAseq signal needs to be plotted}

\item{nb_clust}{An interger specifying the number of cluster}

\item{pf_col_label}{A platform column name use to label gene axis}

\item{exp_grp_col}{An exp_grp column name use to label sample}

\item{exp_grp_col_label}{exprimental}

\item{second}{exprimental}

\item{BREAK_TOO_EXPENSIVE}{A boolean specifying if too expensive operation must be breaked.}

\item{...}{argument passed to plot_rna_sig function}

\item{study}{An epimedtools Study RC object}
}
\description{
This function runs heatmap app.
}

