% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fake_study}
\alias{get_fake_study}
\title{A Function That Builds a Fake Study.}
\usage{
get_fake_study(nb_samples = 12, nb_probes = 10)
}
\arguments{
\item{nb_samples}{An integer that describes the number of samples.}

\item{nb_probes}{An integer that describes the number of probes.}
}
\value{
a fake study.
}
\description{
This function builds a fake study.
}

