% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_geo.R
\name{get_gsm}
\alias{get_gsm}
\title{Retrieving .CEL.gz Files from GEO.}
\usage{
get_gsm(gsms, basedir)
}
\arguments{
\item{gsms}{A character vector of GMSs to be retrived.}

\item{basedir}{A string describing the targeted directory.}
}
\value{
A character vector of .CEL.gz files.
}
\description{
This function retrieves .CEL.gz files from GEO to the file system. It takes as parameters a character vector of GMSs to be retrived and a targeted directory. It retruns a character vector of .CEL.gz files.
}

