% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nbline}
\alias{get_nbline}
\title{A Function That Computes the Number of Line a a Text File}
\usage{
get_nbline(x)
}
\arguments{
\item{x}{a string specifying the text file}
}
\description{
This function returns the number of line a a text fileplots heatmaps.
}

