% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tcga_exp_grp}
\alias{get_tcga_exp_grp}
\title{Retrieve exp_grp of a given TCGA project from epimeddb}
\usage{
get_tcga_exp_grp(tcga_project)
}
\arguments{
\item{tcga_project}{the name of the TCGA project to retrieve}
}
\description{
This function retrieves exp_grp of a given TCGA project from epimeddb
}
