% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transcriptogram}
\alias{get_transcriptogram}
\title{A Function That Computes Transcriptogram of a Gene.}
\usage{
get_transcriptogram(exons_bed, coverage_bwfiles, bin_size = 5,
  tmp_dir = "tmp")
}
\arguments{
\item{exons_bed}{a bed-like dataframe of exons.}

\item{coverage_bwfiles}{a vector of bigwig files.}

\item{bin_size}{an integer specifying the bin to discretize RNA-Seq signal.}

\item{tmp_dir}{a string specifying temporary directory.}
}
\description{
This function computes  RNA-Seq signal matrix of a gene from coverage signal.
}

