% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersect_rec}
\alias{intersect_rec}
\title{A Function That Perform Intersection Between Sets.}
\usage{
intersect_rec(x, y, ...)
}
\arguments{
\item{x}{set: vectors (of the same mode) containing a sequence of items (conceptually) with no duplicated values.}

\item{y}{set: vectors (of the same mode) containing a sequence of items (conceptually) with no duplicated values.}

\item{...}{set: vectors (of the same mode) containing a sequence of items (conceptually) with no duplicated values.}
}
\description{
This function use recurssively intersect function
}
