% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longest_common_prefix}
\alias{longest_common_prefix}
\title{A Function That Returns the Longest Common Prefix.}
\usage{
longest_common_prefix(s, split = "", fixed = FALSE)
}
\arguments{
\item{s}{A vector of character.}

\item{split}{character passed to strsplit function as split.}

\item{fixed}{boolean passed to gsub function as fixed.}
}
\description{
This function returns the longest common prefix.
}
