% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m2sd}
\alias{m2sd}
\title{A Function That Computes `mean` + 2 * `sd` on a Numeric Vector.}
\usage{
m2sd(ctrl)
}
\arguments{
\item{ctrl}{A numeric vector}
}
\value{
`mean` + 2 * `sd` of the inpuit vector
}
\description{
This function compute `mean` + 2 * `sd` on a numeric vector.
}

