% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgetGEO}
\alias{mgetGEO}
\title{A memoised version of getGEO}
\usage{
mgetGEO(...)
}
\arguments{
\item{...}{parameters passed to getGEO function.}
}
\value{
GEO content.
}
\description{
This fucntion offer a memoised version of getGEO. It is used to get GEO data 
from GEO portal or a local file.
}
\examples{
# foo = mgetGEO("GSE42707", getGPL=FALSE)
}

