% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{monitored_apply}
\alias{monitored_apply}
\title{A Monitored Version of `apply`}
\usage{
monitored_apply(mat, marg = 1, func, mod = 100, ...)
}
\arguments{
\item{mat}{A matrix usualy gives as `apply` first parameter.}

\item{marg}{An integer describing the marginal to use, usualy gives as  `apply` second parameter.}

\item{func}{A function usualy gives as `sapply` third parameter.}

\item{mod}{An integer that define the frequency of the monitoring.}

\item{...}{Parameters passed to `func` function.}
}
\value{
A vector of the application of the `func` function to each element of the `vec` vector.
}
\description{
This function offer a monitored version of `apply`.
}
