% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mread.csv2}
\alias{mread.csv2}
\title{A memoised version of read.csv2}
\usage{
mread.csv2(...)
}
\arguments{
\item{...}{parameters passed to read.csv2 function.}
}
\value{
csv file content.
}
\description{
This function offer a memoised version of read.csv2.
}
