% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mreadRDS}
\alias{mreadRDS}
\title{A memoised version of readRDS}
\usage{
mreadRDS(...)
}
\arguments{
\item{...}{parameters passed to readRDS function.}
}
\value{
RDS content.
}
\description{
This function offer a memoised version of readRDS. It is used to get GEO data
from GEO portal or a local file.
}
\examples{
# foo = mreadRDS("GSE42707", getGPL=FALSE)
}

