% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mtgetGEO}
\alias{mtgetGEO}
\title{An other memoised version of getGEO}
\usage{
mtgetGEO(...)
}
\arguments{
\item{...}{parameters passed to getGEO function.}
}
\value{
GEO content.
}
\description{
This fucntion offer a memoised version of getGEO. It returns a Table 
version of a GEO query response. It is used to get platform metadatas 
from GEO portal or a local file.
}
\examples{
# foo = mgetGEO("GPL13534", getGPL=FALSE)
}

