% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overview}
\alias{overview}
\title{A Function That Overviews an Experiment Grouping}
\usage{
overview(e, FULL = FALSE)
}
\arguments{
\item{e}{An expriment grouping}

\item{FULL}{A boolean to specifying if homogenous or fully heterogenous colums need to be reported}
}
\value{
Nothing
}
\description{
This function overviews an experiment grouping.
}

