% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform_anova_gen}
\alias{perform_anova_gen}
\title{A Function Performs Differential Expression Analysis Based On ANOVA}
\usage{
perform_anova_gen(design, model, data, key = NULL, MONITORED_APPLY = FALSE,
  AD_TEST = TRUE)
}
\arguments{
\item{design}{A dataframe that describes the design of the experiment.}

\item{model}{A model describe the test to apply.}

\item{data}{A matrix of exrpression values for probe/gene (lines) and confditions (colums).}

\item{key}{A character string that will suffix the column names of the resulting data frame}

\item{MONITORED_APPLY}{A boolean set to TRUE if we want to monitor the loop on lines (usefull for debugging).}

\item{AD_TEST}{A boolean set to TRUE if we want to perform normality test.}
}
\description{
This function peforms an ANOVA for each probe/gene (lines) across confditions (colums) of a data matrix. 
Its return a dataframe that includes many metrics as a results (beta, pval...).
Genes are indexed by `g` in `G`.
Conditions are indexed by `h` in `H` (histology, sex...).
Samples are indexed by `i` in `I`.
For each gene `g`:
}
\details{
* we compute a linear model of the level of expression value according to histological groups: `log2(exprs_{g, i}) = beta_{g, h} + epsilon_{g, i}`.
  * we perform ANOVA test on this model and harvest corresponding p-value
  * we extract of each group: `beta_{g, h} = frac{sum_{g in G} mean(log2(exprs_{g,i}))}{|G|} - mean(log2(exprs_{g,i}))`
  * we compute corresponding logratio scores: `logratio_{g, h} = (1 + frac{1}/{|H| - 1}) beta_{g, h}`
  * we compute corresponding foldchange scores using `gtools::foldchange2logratio` function

This value are reported into the `anova_lung_res.csv` matrix file which fields are:

* probename:            the name of the corresponding probe      
  * ad_pval:              the result of the Anderson-Darling normality test (hypothesis for ANOVA)  
  * intercept:            the mean of means of each group 
  * beta_histo_ADC:       the previously describe beta value for the group `ADC` of the experimental grouping field `histo`
  * logratio_histo_ADC :  the previously describe logratio value for the group `ADC` of the experimental grouping field `histo`   
  * foldchange_histo_ADC: the previously describe foldchange value for the group `ADC` of the experimental grouping field `histo`
  * pval_histo:           the p-value associated to the ANOVA test   
  * adj_pval_histo:       the adjusted p-value associated to the ANOVA test using Benjamini-Hochberg procedure
  * lower_gs:             the name of the associated gene in lower case
}

