% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_hm2}
\alias{plot_hm2}
\title{A Function That Plots Heatmaps.}
\usage{
plot_hm2(study, pf_col, exp_grp_col_label, exp_grp_idx, platform_idx,
  method_hclust = "complete", nb_clust = 2, var = "raw",
  PLOT_HM_raw = TRUE, USE_CLUST = FALSE, CLUST_ROW = FALSE,
  CLUST_COL = FALSE, BREAK_TOO_EXPENSIVE = TRUE,
  colors = c("royalblue", "springgreen", "yellow", "red"), ...)
}
\arguments{
\item{study}{An epimedtools Study RC object}

\item{pf_col}{A platform column name use to label gene axis}

\item{exp_grp_col_label}{An exp_grp column name use to label sample}

\item{exp_grp_idx}{A vector indexing samples}

\item{platform_idx}{A vector indexing genes}

\item{method_hclust}{A string specifying hclust method to use}

\item{nb_clust}{An interger specifying the number of cluster}

\item{var}{A string defining the distance used to cluster data}

\item{PLOT_HM_raw}{A boolean specifying if heatmap needs to be ploted}

\item{USE_CLUST}{A boolean specifying if dendogram needs to be ploted}

\item{CLUST_ROW}{A boolean specifying if row dendogram needs to be ploted}

\item{CLUST_COL}{A boolean specifying if column dendogram needs to be ploted}

\item{BREAK_TOO_EXPENSIVE}{A boolean specifying if too expensive operation must be breaked.}

\item{colors}{Colors to interpolate; must be a valid argument to col2rgb().}

\item{...}{Parameters passed to gplots::heatmap.2 function}
}
\description{
This function plots heatmaps.
}
