% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_survival_panel}
\alias{plot_survival_panel}
\title{A Function That Plots Survival Panel.}
\usage{
plot_survival_panel(probe_name, sample_names, exp_grp_key, study, nb_q = 5,
  anova_mw_res, gene_pf_colname)
}
\arguments{
\item{probe_name}{probe to consider}

\item{sample_names}{index of samples to consider}

\item{exp_grp_key}{the column name of the exprimental grouping to considere}

\item{study}{An Reference classes object that contains data and metadata}

\item{nb_q}{An integer specifying the number of quantile to initialize discretized expression data}

\item{anova_mw_res}{A dataframe with ANOVA and Mann-Whitney results}

\item{gene_pf_colname}{A character string specifying the name of the platform column to use for the gene name.}
}
\description{
This function plots ...
}

