% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_survival_panel_simple}
\alias{plot_survival_panel_simple}
\title{A Function That Plots Survival Panel}
\usage{
plot_survival_panel_simple(probe_name, sample_names, study, nb_q = 5,
  gene_pf_colname = "lower_gs", ss_key = "os",
  colors = c("deepskyblue", "black", "red"), main, ...)
}
\arguments{
\item{probe_name}{probe to consider}

\item{sample_names}{index of samples to consider}

\item{study}{An Reference classes object that contains data and metadata}

\item{nb_q}{An integer specifying the number of quantile to initialize discretized expression data}

\item{gene_pf_colname}{A character string specifying the name of the platform column to use for the gene name}

\item{ss_key}{A character string specifying the experimental grouping column to use for survival}

\item{colors}{Colors to interpolate; must be a valid argument to col2rgb().}

\item{main}{A character string to explicit the title of the plot}

\item{...}{Parameters passed to plot_survival_panel_simple2 function.}
}
\description{
This function plots ...
}
