% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_survival_panel_simple2}
\alias{plot_survival_panel_simple2}
\title{A Function That Plots Survival Panel}
\usage{
plot_survival_panel_simple2(ss, v, nb_q = 5,
  gene_pf_colname = "lower_gs", colors = c("deepskyblue", "black",
  "red"), main, thresh, PLOT = TRUE, ...)
}
\arguments{
\item{ss}{A survival structure such as produced by function Surv of package survival.}

\item{v}{A (discretized) vector indexed as ss.}

\item{nb_q}{An integer specifying the number of quantile to initialize discretized expression data}

\item{gene_pf_colname}{A character string specifying the name of the platform column to use for the gene name}

\item{colors}{Colors to interpolate; must be a valid argument to col2rgb()}

\item{main}{A character string to explicit the title of the plot}

\item{thresh}{A vector of integer used fopr thresholds}

\item{PLOT}{A boolean set to TRUE if data needs to be plotted.}

\item{...}{Parameters passed to scurve function}
}
\description{
This function plots ...
}
