% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_volcano}
\alias{plot_volcano}
\title{A Function That Draws a Volcano Plot}
\usage{
plot_volcano(res_key_lr, res_key_pval, anova_mw_res, fc_thres = c(-1.5,
  -1.2, 1.2, 1.5), pval_thres = 0.05, legend_place = "bottomright",
  LEGEND = TRUE, ...)
}
\arguments{
\item{res_key_lr}{the column name of the result dataframe to considere as logratio}

\item{res_key_pval}{the column name of the result dataframe to considere as logratio}

\item{anova_mw_res}{A dataframe with ANOVA and Mann-Whitney results}

\item{fc_thres}{A vector of integer corresponding to the foldchange thresholds}

\item{pval_thres}{An integer corresponding to the p-value thresholds}

\item{legend_place}{A character string to specify where to put the legend}

\item{LEGEND}{A boolean to specifying if legend need to be plotted}

\item{...}{Parameters passed to plot function}
}
\description{
This function draws volcano plot, logratio vs. p-value.
}
