% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{req}
\alias{req}
\title{A Function That Requests Epimed Database}
\usage{
req(query, dbname = "epimed_prod", host = "epimed-db.imag.fr",
  port = 5432, user = "epimedtools", password = "epimedtools")
}
\arguments{
\item{query}{A character string indicating to SQL query to execute}

\item{dbname}{A character string integer the nane of the database}

\item{host}{A character string the host of the database}

\item{port}{An integer indicating the port for the connexion to the database}

\item{user}{A character string indicating the user tuse for the connexion to the database}

\item{password}{A character string indicating the password to use for the connexion to the database}
}
\value{
a dataframe corresponding to data requested
}
\description{
This function requests epimed database.
}

