% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revcomp}
\alias{revcomp}
\title{A Function That Computes the Reverse Complement of a DNA Sequence}
\usage{
revcomp(x)
}
\arguments{
\item{x}{a string specifying the DNA sequence}
}
\description{
This function returns the Reverse complement of a DNA sequence.
}
