% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scurve}
\alias{scurve}
\title{A Function That Plots Survival Curves.}
\usage{
scurve(SS, v, colors = c("deepskyblue", "black", "red"), main = "Survival",
  legend, ...)
}
\arguments{
\item{SS}{A survival structure such as produced by function Surv of package survival.}

\item{v}{A (discretized) vector indexed as SS.}

\item{colors}{Colors to interpolate; must be a valid argument to col2rgb().}

\item{main}{A character string to explicit the title of the plot}

\item{legend}{A vector of character to explicit the legend of the plot}

\item{...}{Parameters passed to plot function.}
}
\description{
This function takes a survival structure such as produced by function Surv of
package survival. Takes a (discretized) vector v, indexed as SS.
Performs the Cox proportional hazard rate
test of SS against v. It plots the survival curves for each
level of v. It also prints the p-value of the log-likelihood test.
}

