% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_column_names}
\alias{simplify_column_names}
\title{A Function That Simplifies Column Names.}
\usage{
simplify_column_names(column_names)
}
\arguments{
\item{column_names}{A character vector that describes the column names.}
}
\value{
A character vector that describes simplified column names.
}
\description{
This function simplyfies column names.
}
