% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_exp_grp}
\alias{simplify_exp_grp}
\title{A Function That Simplifies Expriment Grouping.}
\usage{
simplify_exp_grp(exp_grp)
}
\arguments{
\item{exp_grp}{a dataframe that describes the experimental grouping.}
}
\value{
a simplified experimental grouping.
}
\description{
This function simplyfies expriment grouping.
}
