% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_factor_names}
\alias{simplify_factor_names}
\title{A Function That Simplifies Factor Names.}
\usage{
simplify_factor_names(factor_names, split = "", fixed = FALSE)
}
\arguments{
\item{factor_names}{A character vector that describes the factor names.}

\item{split}{character passed to strsplit function as split.}

\item{fixed}{boolean passed to gsub function as fixed.}
}
\value{
A character vector that describes simplified factor names.
}
\description{
This function simplyfies factor names.
}

