% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_sample_names}
\alias{simplify_sample_names}
\title{A Function That Simplify Sample Names.}
\usage{
simplify_sample_names(sample_names)
}
\arguments{
\item{sample_names}{A character vector that describes the sample names.}
}
\value{
A character vector that describes simplified sample names.
}
\description{
This function simplfies sample names.
}

