% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study-class}
\alias{Study}
\alias{Study-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class allow to manipulate useful concepts involved in a multi-omics
study.
}
\section{Fields}{

\describe{
\item{\code{gset}}{The GEO set of data as it is return by getGEO.}

\item{\code{plaform}}{The Table version of the GEO platform description.}

\item{\code{gse}}{A character string that describes the GEO accession number.}

\item{\code{platform_filenames}}{A character string that describes the platform file name.}

\item{\code{series_matrix_filename}}{A character string that describes the GEO serie matrix family file name.}

\item{\code{cache_filename}}{A character string that describes the study cache file name.}
}}
\section{Methods}{

\describe{
\item{\code{cache_it(cache_filename)}}{Writes the study on the disk.}

\item{\code{get_data(...)}}{Proxy to data hold in gset.}

\item{\code{get_exp_grp(...)}}{Proxy to phenoData@data hold in gset.}

\item{\code{get_gset(CACHE = TRUE, MEMOISE = FALSE, dest_dir = "../../data")}}{Computes (if not yet done) and returns the gset field.}

\item{\code{get_platform(CACHE = FALSE, MEMOISE = TRUE,
  dest_dir = "../../data/platforms")}}{Computes if not and returns the platform field.}

\item{\code{get_platform_name(...)}}{Proxy to annotation hold in gset.}

\item{\code{initialize(cfn)}}{Constructor.}

\item{\code{plot_qc(method = "boxplot", ...)}}{Plot the quality control of the study.}
}}

