% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study_abstract-class}
\alias{Study_abstract}
\alias{Study_abstract-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class allow to manipulate useful concepts involved in a multi-omics
study. It an abstract class.
}
\section{Fields}{

\describe{
\item{\code{plaform}}{The Table version of the GEO platform description.}

\item{\code{platform_filenames}}{A character string that describes the platform file name.}

\item{\code{cache_filename}}{A character string that describes the study cache file name.}
}}
\section{Methods}{

\describe{
\item{\code{cache_it(cache_filename)}}{Writes the study on the disk.}

\item{\code{get_data(...)}}{Abstract method that gives access to the data.}

\item{\code{get_exp_grp(...)}}{Abstract method that gives access to the experimental grouping.}

\item{\code{get_platform(CACHE = FALSE, MEMOISE = TRUE,
  dest_dir = "../../data/platforms")}}{Computes if not and returns the platform field.}

\item{\code{get_platform_name(...)}}{Abstract method that gives access to the platform name.}

\item{\code{initialize(cfn)}}{Constructor.}

\item{\code{plot_qc(method = "boxplot", ...)}}{Plot the quality control of the study.}
}}

