% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study_geo-class}
\alias{Study_geo}
\alias{Study_geo-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class extends Study_abstract by encapsulation GEO dataset.
}
\section{Fields}{

\describe{
\item{\code{gset}}{The GEO set of data as it is return by getGEO.}

\item{\code{gse}}{A character string that describes the GEO accession number.}

\item{\code{series_matrix_filename}}{A character string that describes the GEO serie matrix family file name.}
}}
\section{Methods}{

\describe{
\item{\code{get_data(...)}}{Abstract method that gives access to the data.}

\item{\code{get_exp_grp(...)}}{Abstract method that gives access to the experimental grouping.}

\item{\code{get_gset(CACHE = TRUE, MEMOISE = FALSE, dest_dir = "../../data")}}{Computes (if not yet done) and returns the gset field.}

\item{\code{get_platform_name(...)}}{Abstract method that gives access to the platform name.}

\item{\code{plot_qc(method = "boxplot", ...)}}{Plot the quality control of the study.}
}}

