% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study_loc-class}
\alias{Study_loc}
\alias{Study_loc-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class extends Study_abstract by encapsulation hand made fields: "data", "platform_name", "exp_grp".
}
\section{Fields}{

\describe{
\item{\code{data}}{A matris of experiemental data. Each column is a sample, named by it's sample name, and each row is a probes.}

\item{\code{platform_name}}{A character string that describes the platform used to perform.}

\item{\code{exp_grp}}{A data.frame that describe the experimental grouping. Each line is a sample.}
}}
\section{Methods}{

\describe{
\item{\code{get_data(...)}}{Abstract method that gives access to the data.}

\item{\code{get_exp_grp(...)}}{Abstract method that gives access to the experimental grouping.}

\item{\code{get_platform_name(...)}}{Abstract method that gives access to the platform name.}
}}

