% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_raw_trscr.R
\docType{class}
\name{Study_raw_trscr-class}
\alias{Study_raw_trscr}
\alias{Study_raw_trscr-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class extends Study_loc by offering methods allowing to build data from cel files.
}
\section{Fields}{

\describe{
\item{\code{cel_filedirs}}{A character string that describes the directory taht contain associated cel files.}
}}
\section{Methods}{

\describe{
\item{\code{set_data(hook, ...)}}{Set the data field and update experiment grouping.}
}}

