% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_exp_grp}
\alias{summarize_exp_grp}
\title{A Function That Summarizes Expriment Grouping.}
\usage{
summarize_exp_grp(exp_grp)
}
\arguments{
\item{exp_grp}{a dataframe that describes the experimental grouping.}
}
\description{
This function summarizes expriment grouping.
}
