% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcga_explore_data}
\alias{tcga_explore_data}
\title{Exploring data}
\usage{
tcga_explore_data(df, lev = 0)
}
\arguments{
\item{df}{data.frame extracted from json}

\item{lev}{level of explorer, increased when recursiv call}
}
\description{
This function explores data
}

