% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcga_fum_data}
\alias{tcga_fum_data}
\title{Fuming data}
\usage{
tcga_fum_data(data, colname)
}
\arguments{
\item{data}{data.frame extracted from json}

\item{colname}{column name to flat.}
}
\description{
This function fums data.
}
