% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coxres}
\alias{coxres}
\title{A Function That Fits the Cox Regression Model}
\usage{
coxres(ss, v)
}
\arguments{
\item{ss}{A survival structure such as produced by function Surv of package survival.}

\item{v}{A (discretized) vector indexed as ss.}

\item{censoring}{date at which data are censored}
}
\value{
A named vector of length 5.
}
\description{
This function takes a survival structure such as produced by function Surv of
package survival. Takes a (discretized) vector v, indexed as ss.
Fits the Cox regression model of ss against v, and tests the
proportional hazards assumption.
Returns as a named vector of length 5:
     pvcox: the significance p-value (likelihood ratio test)
     pvhz: the p-value for the validity of the model
     hrlb: the lower bound of the 95% confidence interval for hr
     hr: the hazard ratio
     hrub: the upper bound of the 95% confidence interval for hr
}
