% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{create_study}
\alias{create_study}
\title{Study Factory}
\usage{
create_study(cache_filename, ...)
}
\arguments{
\item{cache_filename}{A character string that describes the study cache file name.}

\item{...}{Parameters that are given to the Study contructor}
}
\value{
It returns an object that extends the Study_abstract class.
}
\description{
Intanciates an object that extends the epimedtools internal class Study_abstract.
}
\details{
This function intanciates an object that extends the epimedtools internal abstract class Study_abstract. You can learn more about Study_abstract class by typing ?epimedtools::Study_abstract
}
\examples{
# create study and load data from GSE57831_series_matrix.txt.gz contained in the package
study = create_study()
study$series_matrix_filename = system.file(
   "extdata/GSE26471", 
   "GSE26471_series_matrix.txt.gz", 
   package = "epimedtools"
 )  
# str(study$get_gset(dest_dir="data"))

# load data from GEO
study = create_study()
study$gse = "GSE26471" # 1 sample, expression
# study$gse = "GSE42707" # 1 sample
# study$gse = "GSE56382" # 2 samples
# study$gse = "GSE49585" # 2 samples
# str(study$get_gset(dest_dir="data"))

# cache study
study$save("/tmp/tmp_cached_study.rds")
# load a cached study
study = create_study("/tmp/tmp_cached_study.rds")
}
