% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_plot_heatmap}
\alias{dt_plot_heatmap}
\title{Binding of deepTools `plotHeatmap` Function.}
\usage{
dt_plot_heatmap(matrix_file, out_filename, y_max, y_min, sort_regions = "no",
  FORCE_EXEC = FALSE)
}
\arguments{
\item{matrix_file}{deepTools `plotHeatmap` parameter.}

\item{out_filename}{deepTools `plotHeatmap` parameter.}

\item{y_max}{deepTools `plotHeatmap` parameter.}

\item{y_min}{deepTools `plotHeatmap` parameter.}

\item{sort_regions}{deepTools `sortRegions` parameter.}

\item{FORCE_EXEC}{a boolean that force call to deepTools `computeMatrix` function.}
}
\description{
This function binds deepTools `plotHeatmap` function.
}

