% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform_anova_gen}
\alias{perform_anova_gen}
\title{A Function Performs Differential Expression Analysis Based On ANOVA}
\usage{
perform_anova_gen(design, model, data, key = NULL,
  MONITORED_APPLY = FALSE, AD_TEST = TRUE)
}
\arguments{
\item{design}{A dataframe that describes the design of the experiment.}

\item{model}{A model describe the test to apply.}

\item{data}{A matrix of exrpression values for probe/gene (lines) and confditions (colums).}

\item{key}{A character string that will suffix the column names of the resulting data frame}

\item{MONITORED_APPLY}{A boolean set to TRUE if we want to monitor the loop on lines (usefull for debugging).}

\item{AD_TEST}{A boolean set to TRUE if we want to perform normality test.}
}
\description{
This function peforms an ANOVA for each probe/gene (lines) across confditions (colums) of a data matrix.
It returns a dataframe that includes many metrics as a results (betas, p-values, ...).
Genes are indexed by `g` in `G`,
conditions are indexed by `h` in `H` (histology, sex...) and
samples are indexed by `i` in `I`.
For each gene `g`:
\itemize{
  \item we compute a linear model of the level of expression value according to histological groups: `log2(exprs_{g, i}) = beta_{g, h} + epsilon_{g, i}`.
  \item we perform ANOVA test on this model and harvest corresponding p-value
  \item we extract of each group: `beta_{g, h} = frac{sum_{g in G} mean(log2(exprs_{g,i}))}{|G|} - mean(log2(exprs_{g,i}))`
  \item we compute corresponding logratio scores: `logratio_{g, h} = (1 + frac{1}/{|H| - 1}) beta_{g, h}`
  \item we compute corresponding foldchange scores using `gtools::foldchange2logratio` function
}
These values are retruned as a dataframe which fields are:
\itemize{
  \item probename:            the name of the corresponding probe
  \item ad_pval:              the result of the Anderson-Darling normality test (hypothesis for ANOVA)
  \item intercept:            the mean of means of each group
  \item beta_hhh_FFF:         the previously describe beta value for the factor `FFF` of the experimental grouping field `hhh`
  \item logratio_hhh_FFF :    the previously describe logratio value for the factor `FFF` of the experimental grouping field `hhh`
  \item foldchange_hhh_FFF:   the previously describe foldchange value for the factor `FFF` of the experimental grouping field `hhh`
  \item pval_hhh:             the p-value associated to the ANOVA test
  \item adj_pval_hhh:         the adjusted p-value associated to the ANOVA test using Benjamini-Hochberg procedure to the threshold of 0.05.
}
}
