% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scurve}
\alias{scurve}
\title{A Function That Plots Survival Curves.}
\usage{
scurve(ss, v, colors = c("deepskyblue", "black", "red"), main = "Survival",
  legend, nb_sign = 3, legend_place = "topright", PLOT_LEGEND = TRUE,
  cex.leg = 1, censoring, ...)
}
\arguments{
\item{ss}{A survival structure such as produced by function Surv of package survival.}

\item{v}{A (discretized) vector indexed as ss.}

\item{colors}{Colors to interpolate; must be a valid argument to col2rgb().}

\item{main}{A character string to explicit the title of the plot}

\item{legend}{A vector of character to explicit the legend of the plot}

\item{nb_sign}{An integer  indicating the number of significant digits to be used}

\item{legend_place}{A character string to specify where to put the legend}

\item{PLOT_LEGEND}{A boolean to specifying if legend need to be plotted}

\item{cex.leg}{A numeric specifying the size of the legend}

\item{censoring}{date at which data are censored}

\item{...}{Parameters passed to plot function.}
}
\description{
This function takes a survival structure such as produced by function Surv of
package survival. Takes a (discretized) vector v, indexed as ss.
Performs the Cox proportional hazard rate
test of ss against v. It plots the survival curves for each
level of v. It also prints the p-value of the log-likelihood test.
}

