% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_factor_names}
\alias{simplify_factor_names}
\title{A Function That Simplify Factor Names.}
\usage{
simplify_factor_names(factor_names)
}
\arguments{
\item{factor_names}{A character vector that describes the factor names.}
}
\value{
A character vector that describes simplified factor names.
}
\description{
This function simplfies factor names.
}

