% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\docType{class}
\name{Study_abstract-class}
\alias{Study_abstract}
\alias{Study_abstract-class}
\title{A Reference Class to represent a multi-omic study.}
\description{
This class allow to manipulate useful concepts involved in a multi-omics
study. It an abstract class.
}
\section{Fields}{

\describe{
\item{\code{plaform}}{The Table version of the GEO platform description.}

\item{\code{platform_filenames}}{A character string that describes the platform file name.}

\item{\code{cache_filename}}{A character string that describes the study cache file name.}

\item{\code{gset}}{The GEO set of data as it is return by getGEO.}

\item{\code{gse}}{A character string that describes the GEO accession number.}

\item{\code{series_matrix_filename}}{A character string that describes the GEO serie matrix family file name.}

\item{\code{data}}{A matris of experiemental data. Each column is a sample, named by it's sample name, and each row is a probes.}

\item{\code{platform_name}}{A character string that describes the platform used to perform.}

\item{\code{exp_grp}}{A data.frame that describe the experimental grouping. Each line is a sample.}
}}
\section{Methods}{

\describe{
\item{\code{do_m2s_analysis(probe_names, exp_grp_key, ctrl_name, nb_perm = 100,
  MONITORED = FALSE, ...)}}{Performs permutation test to detect right shifted exprtession groups for a given `probe_name` and a given `factor_name`.}

\item{\code{get_cel_files(dest_dir = "data", ...)}}{Retrieve .CEL.gz from NCBI GEO eweb site}

\item{\code{get_gset(CACHE = TRUE, MEMOISE = FALSE, dest_dir = "data")}}{Computes (if not yet done) and returns the gset field.}

\item{\code{get_platform(CACHE = FALSE, MEMOISE = TRUE, dest_dir = "data/platforms")}}{Computes if not and returns the platform field.}

\item{\code{get_probe_gene_tab(gene, DEEP_SEARCH = FALSE, pf_col_name = "Gene Symbol",
  ...)}}{Build a gene / probe table from a gene list.}

\item{\code{get_ratio(ctrl_sample_names, method = "mean", ...)}}{Normalize data accross probes using method *method* according to a reference poulation describe by  *ctrl_sample_names* its sample names vector.}

\item{\code{gs_to_probe(gene, ALL_PF_COL = FALSE, pf_col_name = "Gene Symbol", ...)}}{Return probes name for a given gene.}

\item{\code{initialize(cfn)}}{Constructor.}

\item{\code{plot_boxplot(probe_name, factor_name, ylim, las = 2, col = "grey",
  border = "black", bp_function_name = "boxplot", ...)}}{Draw the box plot for a given `probe_name` and a given `factor_name`.}

\item{\code{plot_qc(method = "boxplot", ...)}}{Plot the quality control of the study.}

\item{\code{save(cache_filename)}}{Writes the study on the disk.}
}}

